mybook/requirements.txt
django == 2.0.3
django-markdownx

